chore: move formatDate to a util file
