+++ frontend/src/components/Note.jsx
@@ -1,16 +1,4 @@
-// Take the datetime format from django and turn it into
-// a user-friendly format
-const formatDate = (isoString) => {
-  const date = new Date(isoString);
-  return date.toLocaleString(undefined, {
-    year: "numeric",
-    month: "short",
-    day: "numeric",
-    hour: "numeric",
-    minute: "2-digit",
-    hour12: true,
-  });
-};
+import formatDate from "../util/formatDate";
 
 const Note = ({ note, handleClick }) => {
   return (
