+++ frontend/src/util/formatDate.js
@@ -1,0 +1,15 @@
+// Take the datetime format from django and turn it into
+// a user-friendly format
+const formatDate = (isoString) => {
+  const date = new Date(isoString);
+  return date.toLocaleString(undefined, {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+    hour: "numeric",
+    minute: "2-digit",
+    hour12: true,
+  });
+};
+
+export default { formatDate };
